% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfregion.R
\name{idfregion}
\alias{idfregion}
\title{Recognize PAC IND ATL ocean region Function
This function allows you to identify location in which ocean region from a numeric vector.}
\usage{
idfregion(lon, lat)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Recognize PAC IND ATL ocean region Function
This function allows you to identify location in which ocean region from a numeric vector.
}
\examples{
idfregion()
}
\keyword{location}
